Move `QueryAnswer` out of reasoner package
