+++ grakn-graql/src/main/java/ai/grakn/graql/internal/query/QueryAnswer.java
@@ -14,9 +14,10 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
+ *
  */
 
-package ai.grakn.graql.internal.reasoner.query;
+package ai.grakn.graql.internal.query;
 
 import ai.grakn.concept.Concept;
 import ai.grakn.graql.VarName;
